Refactor license adoption rules into separate classes
Refactored `LicenseChoicesFactory` by splitting license adoption logic into dedicated classes for better modularity and readability. Introduced `LicenseIdAdoptionRule`, `OtherLicenseAdoptionRule`, and `SPDXExpressionAdoptionRule` to handle specific scenarios, with a shared base class `AdoptionLicenseRuleBase`. Enhanced logging in `ClearlyDefinedClient` for missing license information.
