+++ src/CdxEnrich/ClearlyDefined/Rules/IAdoptionLicenseRule.cs
@@ -1,0 +1,11 @@
+﻿using CycloneDX.Models;
+using PackageUrl;
+
+namespace CdxEnrich.ClearlyDefined.Rules
+{
+    public interface IAdoptionLicenseRule
+    {
+        bool CanApply(ClearlyDefinedResponse.LicensedData dataLicensed);
+        List<LicenseChoice>? Apply(PackageURL packageUrl, ClearlyDefinedResponse.LicensedData dataLicensed);
+    }
+}
